feat: PTY terminal infrastructure with job control, PS/2 keyboard, and FreeBSD ABI compatibility
Implement a modular TTY stack replacing the monolithic console:
Console (HW input) → TTYD (line discipline) → VFS (/dev/pts/N) → shell

Kernel:
- I/O APIC initialization: route IRQ1 (keyboard) and IRQ4 (serial) to BSP
- Automatic user stack growth on page fault (1 MiB window below RSP)
- Async NBSend 64-message ring buffer for fire-and-forget IPC
- PS/2 keyboard capability slots (IoPort 0x60-0x64, IRQ1 handler)
- Notification capability badging (previously endpoint-only)
- Defensive IPC buffer validation against unmapped pages

Userland:
- ttyd: PTY daemon with full line discipline (ICANON, ISIG, ECHO,
  OPOST/ONLCR), per-PTY termios, Ctrl-C/Z/\ signal generation to
  foreground process group, 4 PTY instances, display TX buffering
- getty: session setup (setsid → open /dev/pts/0 → TIOCSCTTY →
  TIOCSPGRP → exec bash), auto-respawned by procmgr on exit
- Console refactored to raw HW input source with PS/2 scan code
  decoding (Set 1), dual IRQ drain (COM1+PS/2), forwarding to ttyd
- VFS: /dev/pts/0 and /dev/ptmx device routing, pending reader queue
  with notification-based wakeup, per-client FD cleanup on exit
- Procmgr: SPAWN_FLAG_RESPAWN auto-restart, PM_KILL_PGID for terminal
  signals, PM_INJECT_CAP for post-spawn capability wiring, session/pgid
  tracking, VFS client exit notification
- Init: shared PTY notification creation, declarative cap injection
  via NeedEP/CopyCap in .service files, bootinfo snapshot frame
- RTLD: graceful exit via procmgr on load failure (replaces panic)

Libraries:
- libsalty: 15 TTYD IPC labels, 5 PM labels, device type constants,
  SPAWN_FLAG_RESPAWN, posix_getsid(), debug tracing infrastructure
- saltyc: FreeBSD-compatible struct stat (timespec fields, st_gen),
  struct dirent (d_namlen), FTS API rewrite, RuneLocale 752-byte
  expansion with case mapping tables, popen()/pclose(), 12 wchar
  classification functions, strtonum(), getbsize(), file locking
  constants, TIOCSCTTY ioctl, sigjmp_buf extension
- Bash port: disable SSE/MMX/FP instructions
- FreeBSD cat: bootstrap patch without Capsicum

+++ lib/saltyc/include/sys/types.h
@@ -2,10 +2,10 @@
 #ifndef __SYS_TYPES_H__
 #define __SYS_TYPES_H__
 
-typedef long           pid_t;
+typedef int            pid_t;
 typedef unsigned int   uid_t;
 typedef unsigned int   gid_t;
-typedef unsigned int   mode_t;
+typedef unsigned short mode_t;
 typedef long           off_t;
 typedef unsigned long  size_t;
 typedef long           ssize_t;
@@ -15,7 +15,7 @@ typedef int            clockid_t;
 typedef unsigned long  ino_t;
 typedef unsigned long  dev_t;
 typedef unsigned long  nlink_t;
-typedef unsigned long  blksize_t;
+typedef int            blksize_t;
 typedef long           blkcnt_t;
 typedef unsigned long  useconds_t;
 typedef long           suseconds_t;
